+++ src/main/java/umc/meme/auth/global/oauth/provider/apple/AppleAuthProvider.java
@@ -1,15 +1,14 @@
-package umc.meme.auth.global.oauth.service.apple;
+package umc.meme.auth.global.oauth.provider.apple;
 
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.security.SignatureException;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
-import umc.meme.auth.domain.user.entity.UserRepository;
 import umc.meme.auth.global.common.status.ErrorStatus;
 import umc.meme.auth.global.exception.AuthException;
 import umc.meme.auth.global.infra.RedisRepository;
-import umc.meme.auth.global.oauth.service.OAuthService;
 import umc.meme.auth.global.oauth.jsonwebkey.PublicKeyDto;
+import umc.meme.auth.global.oauth.provider.OAuthProvider;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -22,7 +21,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
-public class AppleAuthService extends OAuthService {
+public class AppleAuthProvider extends OAuthProvider {
 
     private static final String REQUEST_URL = "https://appleid.apple.com/auth/keys";
     private static final String PROVIDER = "APPLE";
@@ -35,7 +34,7 @@ public class AppleAuthService extends OAuthService {
 
     private final RedisRepository redisRepository;
 
-    public AppleAuthService(RedisRepository redisRepository) {
+    public AppleAuthProvider(RedisRepository redisRepository) {
         this.redisRepository = redisRepository;
     }
 
@@ -97,7 +96,6 @@ public class AppleAuthService extends OAuthService {
                     .key(jsonData)
                     .build());
         } else {
-            System.out.println("RESPONSE_CODE = " + responseCode);
             throw new IOException();
         }
     }
